AudioBugfix for Samsung/Android Browser
On Samsung Devices (on my Galaxy S4Mini) playing a sound effect
sometimes results in a crash, throwing the following Exception:

InvalidStateError: DOM Exception 11

I've found out that it could happen that chan.waSource.stop( 0 )
throws this exception. However, as far as I understand the source,
this happens if stop is called twice or when waSource Object
isn't playing. So this if-branch shouldn't be reached, but it does.

I've fixed it by catching the exception and ignoring it, which is
fine in this case.

I'm sure there must be a more cleaner solution, but I wasn't
able to find one. This fix worked fine for me.